+++ Game/AI/Decks/KashtiraExecutor.cs
@@ -224,7 +224,7 @@ namespace WindBot.Game.AI.Decks
             }
             return base.OnSelectPlace(cardId, player, location, available);
         }
-        public override IList<ClientCard> OnSelectCard(IList<ClientCard> cards, int min, int max, long hint, bool cancelable)
+        public override IList<ClientCard> OnSelectCard(IList<ClientCard> cards, int min, int max, int hint, bool cancelable)
         {
             if (cards.Any(card => card != null && card.Location == CardLocation.Extra
                  && hint == HintMsg.Remove && min == 1 && max == 1))
