+++ Game/AI/Decks/[ImaginaryArkExecutor].cs
@@ -139,7 +139,7 @@ namespace WindBot.Game.AI.Decks
             IList<ClientCard> cards,
             int min,
             int max,
-            long hint,
+            int hint,
             bool cancelable
         )
         {
