+++ Game/GameBehavior.cs
@@ -26,7 +26,7 @@ namespace WindBot.Game
         private Duel _duel;
         private int _hand;
         private bool _debug;
-        private long _select_hint;
+        private int _select_hint;
         private GameMessage _lastMessage;
         public class LocationInfo
         {
@@ -390,7 +390,7 @@ namespace WindBot.Game
         {
             int type = packet.ReadByte();
             int player = packet.ReadByte();
-            long data = packet.ReadInt64();
+            int data = packet.ReadInt32();
             if (type == 1) // HINT_EVENT
             {
                 if (data == 24) // battling
@@ -1020,7 +1020,7 @@ namespace WindBot.Game
             Connection.Send(CtosMessage.Response, _ai.OnSelectBattleCmd(battle).ToValue());
         }
 
-        private void InternalOnSelectCard(BinaryReader packet, Func<IList<ClientCard>, int, int, long, bool, IList<ClientCard>> func, bool tribute = false)
+        private void InternalOnSelectCard(BinaryReader packet, Func<IList<ClientCard>, int, int, int, bool, IList<ClientCard>> func, bool tribute = false)
         {
             int player = packet.ReadByte();
             bool cancelable = packet.ReadByte() != 0;
@@ -1086,7 +1086,7 @@ namespace WindBot.Game
             Connection.Send(reply);
         }
 
-        private void InternalOnSelectUnselectCard(BinaryReader packet, Func<IList<ClientCard>, int, int, long, bool, IList<ClientCard>> func)
+        private void InternalOnSelectUnselectCard(BinaryReader packet, Func<IList<ClientCard>, int, int, int, bool, IList<ClientCard>> func)
         {
             int player = packet.ReadByte();
             bool finishable = packet.ReadByte() != 0;
